flows: load messages dynamically upon card refresh
