+++ .pylintrc
@@ -5,4 +5,4 @@ load-plugins=pylint_django,pylint.extensions.bad_builtin
 extension-pkg-whitelist=lxml
 const-rgx=[a-zA-Z0-9_]{1,40}$
 ignored-modules=django-otp
-jobs=4
+jobs=12
