+++ passbook/flows/views.py
@@ -187,4 +187,5 @@ class FlowExecutorShellView(TemplateView):
 
     def get_context_data(self, **kwargs) -> Dict[str, Any]:
         kwargs["exec_url"] = reverse("passbook_flows:flow-executor", kwargs=self.kwargs)
+        kwargs["msg_url"] = reverse("passbook_api:messages-list")
         return kwargs
